+++ src/components/TagSet/tagSetChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "New Feature",
+    affects: ["Documentation", "Functionality", "Styles"],
+    notes: ["Adds semantic color variants."],
+  },
+  {
     date: "2025-08-11",
     version: "4.0.0",
     type: "Update",