+++ CHANGELOG.md
@@ -8,6 +8,15 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Updates
+
+- Updates the `TagSet` component to add semantic color variants.
+- Updates the hex color value for `ui.warning.tertiary` for better accessibility compliance.
+
+### Fixes
+
+- Fixes the hex color value for `dark.ui.warning.secondary`.
+
 ## 4.1.2 (January 29, 2026)
 
 ### Updates